re gen GetStatusesLookupRequestV1
